# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mobile Dashboard API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/billing/generate": {
            "post": {
                "tags": [
                    "Billing"
                ],
                "summary": "Generate monthly billing",
                "operationId": "5725bfb33efad0b9d9db71e59466d411",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "entity_id": {
                                        "type": "integer"
                                    },
                                    "plant_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "month": {
                                        "type": "string",
                                        "example": "2026-05"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Billing generated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/billing/history": {
            "post": {
                "tags": [
                    "Billing"
                ],
                "summary": "Get billing history",
                "operationId": "fa9978d3723c0661ad0e50677e9ebeb2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "entity_id": {
                                        "type": "integer"
                                    },
                                    "plant_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Billing history"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/billing/{billing}/pay": {
            "post": {
                "tags": [
                    "Billing"
                ],
                "summary": "Mock payment for billing",
                "operationId": "e7b7d67ca6149613acc81fbbd02fe682",
                "parameters": [
                    {
                        "name": "billing",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "entity_id": {
                                        "type": "integer"
                                    },
                                    "plant_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "success": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment status updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/dashboard/sales-summary": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get sales summary (Donut Chart data)",
                "operationId": "ac2048c55cd641389f0487fab35cb150",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/sales-stats": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get sales statistics (MT, UNIT/CFT, TRIPS)",
                "operationId": "81c84b5f1e3f50c39cd17355ec26932f",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/top-products": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get top selling products",
                "operationId": "b45446e18fe41f2c035f6cccdd9a48a9",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/sales-details": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get detailed sales breakdown (Cash vs Credit)",
                "operationId": "516b930161da0a498c749cd4cb6e569f",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/stock-details": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get current stock levels",
                "operationId": "b161cede51b5418dcb503927bca1fc0d",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/trips-details": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get trip counts by vehicle",
                "operationId": "e0ce12fc3e3e3735ea08510530eacbd4",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/customer-details": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get recent customer activities",
                "operationId": "80ee7e52aaf79ad87bc8ac87daf36f9f",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/alerts": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard alerts",
                "operationId": "9242cc94b1a654b1ae06b366db8937d7",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/master/plants": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Get plant list for filtering",
                "operationId": "64b1803a036836815837800251e20aac",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}